重命名间隔可以设置为0.5分钟

+++ src/main/java/ani/rss/entity/Config.java
@@ -73,7 +73,7 @@ public class Config implements Serializable {
     /**
      * 重命名间隔(分钟)
      */
-    private Integer renameSleep;
+    private Double renameSleep;
 
     /**
      * 自动重命名
